+++ src/se/liu/danal315samak519/entities/enemies/Knight.java
@@ -8,7 +8,7 @@ import java.awt.geom.Point2D;
 public class Knight extends Enemy
 {
 
-    public Knight(final Point2D.Double coord, final Player player) {
+    public Knight(final Point2D.Float coord, final Player player) {
 	super(coord, player);
 	setMaxSpeed(2);
 	setStats(5, 2);
