+++ src/se/liu/danal315samak519/weapons/Sword.java
@@ -6,24 +6,24 @@ import java.awt.geom.Point2D;
 
 public class Sword extends Weapon
 {
-    public Sword(final Point2D.Double coord, final Character owner) {
+    public Sword(final Point2D.Float coord, final Character owner) {
 	super(coord, owner);
 	switch (owner.getDir()) {
 	    case UP:
 		setSize(10, 60);
-		setLocation((owner.getX() + (owner.getWidth() / 2.0) - this.getWidth() / 2.0), (owner.getY() - this.getHeight()));
+		setLocation((owner.getX() + (owner.getWidth() / 2.0f) - this.getWidth() / 2.0f), (owner.getY() - this.getHeight()));
 		break;
 	    case DOWN:
-		setSize(10, 60);
-		setLocation((owner.getX() + (owner.getWidth() / 2.0) - this.getWidth() / 2.0), (owner.getY() + (owner.getHeight())));
+		setSize(10f, 60f);
+		setLocation((owner.getX() + (owner.getWidth() / 2.0f) - this.getWidth() / 2.0f), (owner.getY() + (owner.getHeight())));
 		break;
 	    case LEFT:
-		setSize(60, 10);
-		setLocation((owner.getX() - this.getWidth()), (owner.getY() + (owner.getHeight() / 2.0)) - (this.getHeight() / 2.0));
+		setSize(60f, 10f);
+		setLocation((owner.getX() - this.getWidth()), (owner.getY() + (owner.getHeight() / 2.0f)) - (this.getHeight() / 2.0f));
 		break;
 	    case RIGHT:
-		setSize(60, 10);
-		setLocation((owner.getX() + owner.getWidth()), (owner.getY() + (owner.getHeight() / 2.0)) - (this.getHeight() / 2.0));
+		setSize(60f, 10f);
+		setLocation((owner.getX() + owner.getWidth()), (owner.getY() + (owner.getHeight() / 2.0f)) - (this.getHeight() / 2.0f));
 		break;
 	}
