+++ src/se/liu/danal315samak519/weapons/Weapon.java
@@ -16,7 +16,7 @@ public abstract class Weapon extends Movable
     protected Character owner;
     protected int lifeSpan;
 
-    protected Weapon(final Point2D.Double coord, final Character owner) {
+    protected Weapon(final Point2D.Float coord, final Character owner) {
 	//We have two choices when it comes to spawning the weapon first look at dir then spawn everything according to that...
 	this.owner = owner;
 	this.lifeSpan = 5;
